% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr.R
\name{corr}
\alias{corr}
\title{Compute correlation matrix}
\usage{
corr(v, normal = F)
}
\arguments{
\item{v}{a data.frame (or list) should be taken.}
}
\description{
Compute Spearman and Pearson correlation matrix.
}
\examples{
corr(mtcars[,1:5])

corr(mtcars[,5:6], normal = T)

}
